+++ clone/clone.go
@@ -1,4 +1,4 @@
-package poly
+package clone
 
 import (
 	"errors"
