+++ sequence/sequence.go
@@ -1,4 +1,4 @@
-package poly
+package sequence
 
 import (
 	"bytes"
